+++ actix/tests/test_address.rs
@@ -4,6 +4,7 @@ extern crate futures;
 extern crate tokio;
 extern crate tokio_timer;
 
+use std::collections::HashSet;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::sync::Arc;
 use std::thread;
@@ -68,7 +69,7 @@ fn test_address() {
             Delay::new(Instant::now() + Duration::new(0, 100)).then(move |_| {
                 addr2.do_send(Ping(4));
 
-                tokio::spawn(Delay::new(Instant::now() + Duration::new(0, 1000)).then(
+                tokio::spawn(Delay::new(Instant::now() + Duration::new(0, 2000)).then(
                     move |_| {
                         System::current().stop();
                         Ok(())
@@ -78,7 +79,7 @@ fn test_address() {
             })
         }));
     });
-    thread::sleep(Duration::from_millis(200));
+    thread::sleep(Duration::from_millis(400));
 
     assert_eq!(count.load(Ordering::Relaxed), 4);
 }
@@ -199,4 +200,117 @@ fn test_call_message_timeout() {
         let _addr2 = TimeoutActor3(addr, count2).start();
     });
     assert_eq!(count.load(Ordering::Relaxed), 1);
+}
+
+#[test]
+fn test_address_eq() {
+    let count0 = Arc::new(AtomicUsize::new(0));
+    let count1 = Arc::clone(&count0);
+
+    System::run(move || {
+        let addr0 = MyActor(count0).start();
+        let addr01 = addr0.clone();
+        let addr02 = addr01.clone();
+
+        assert!(addr0 == addr01);
+        assert!(addr0 == addr02);
+
+        let addr1 = MyActor(count1).start();
+
+        assert!(addr0 != addr1);
+
+        System::current().stop();
+    });
+}
+
+#[test]
+fn test_address_hash() {
+    let count0 = Arc::new(AtomicUsize::new(0));
+    let count1 = Arc::clone(&count0);
+
+    System::run(move || {
+        let addr0 = MyActor(count0).start();
+        let addr01 = addr0.clone();
+
+        let mut addresses = HashSet::new();
+        addresses.insert(addr0.clone());
+        addresses.insert(addr01.clone());
+
+        assert_eq!(addresses.len(), 1);
+        assert!(addresses.contains(&addr0));
+        assert!(addresses.contains(&addr01));
+
+        let addr1 = MyActor(count1).start();
+        addresses.insert(addr1.clone());
+
+        assert_eq!(addresses.len(), 2);
+        assert!(addresses.contains(&addr1));
+
+        assert!(addresses.remove(&addr0));
+        assert!(!addresses.contains(&addr0));
+        assert!(!addresses.contains(&addr01));
+        assert_eq!(addresses.len(), 1);
+        assert!(addresses.contains(&addr1));
+
+        System::current().stop();
+    });
+}
+
+#[test]
+fn test_recipient_eq() {
+    let count0 = Arc::new(AtomicUsize::new(0));
+    let count1 = Arc::clone(&count0);
+
+    System::run(move || {
+        let addr0 = MyActor(count0).start();
+        let recipient01 = addr0.clone().recipient::<Ping>();
+        let recipient02 = addr0.clone().recipient::<Ping>();
+
+        assert!(recipient01 == recipient02);
+
+        let recipient03 = recipient01.clone();
+        assert!(recipient01 == recipient03);
+
+        let addr1 = MyActor(count1).start();
+        let recipient11 = addr1.clone().recipient::<Ping>();
+
+        assert!(recipient01 != recipient11);
+
+        System::current().stop();
+    });
+}
+
+#[test]
+fn test_recipient_hash() {
+    let count0 = Arc::new(AtomicUsize::new(0));
+    let count1 = Arc::clone(&count0);
+
+    System::run(move || {
+        let addr0 = MyActor(count0).start();
+        let recipient01 = addr0.clone().recipient::<Ping>();
+        let recipient02 = addr0.clone().recipient::<Ping>();
+
+        let mut recipients = HashSet::new();
+        recipients.insert(recipient01.clone());
+        recipients.insert(recipient02.clone());
+
+        assert_eq!(recipients.len(), 1);
+        assert!(recipients.contains(&recipient01));
+        assert!(recipients.contains(&recipient02));
+
+        let addr1 = MyActor(count1).start();
+        let recipient11 = addr1.clone().recipient::<Ping>();
+        recipients.insert(recipient11.clone());
+
+        assert_eq!(recipients.len(), 2);
+        assert!(recipients.contains(&recipient11));
+
+        assert!(recipients.remove(&recipient01));
+        assert!(!recipients.contains(&recipient01));
+        assert!(!recipients.contains(&recipient02));
+        assert_eq!(recipients.len(), 1);
+        assert!(recipients.contains(&recipient11));
+
+        System::current().stop();
+    });
 }